Implement a soft timeout
As the comments states, the Unicorn timeout mechanism takes no prisoner.
When the deadline is hit, the master process send SIGKILL to the worker.

This is good because it's extremely reliable. Even if the process is deeply
deadlocked, SIGKILL should make it exit.

However, in the vast majority of cases, such heavy handed technique isn't
necessary, and handling the timeout inside the master process means we
cannot extract much debugging information to fix the root cause.

As such it's preferable to give the opportunity to the worker to enforce
its own timeout, so that when it happens it can phone back home and
alert the service maintainer. If somehow the worker isn't able to
enforce its own timeout, the master process can send a SIGKILL.

Additionally, a single timeout value for the entire service is hard to
enforce, inevitably one endpoint requires a much longer timeout,
forcing the global timeout to be raised.

Having a soft timeout in the worker allows to selectively push back
the deadline based on application logic, for instance relaxing the
timeout for a single API endpoint, keeping it low for the rest of
the service.